[Test] 통합테스트 진행중

+++ src/main/java/org/example/ticketing/api/usecase/reservation/MakeReservationUseCase.java
@@ -1,5 +1,6 @@
 package org.example.ticketing.api.usecase.reservation;
 
+import jakarta.transaction.Transactional;
 import org.example.ticketing.api.dto.request.ReservationRequestDTO;
 import org.example.ticketing.api.dto.request.UserRequestDTO;
 import org.example.ticketing.api.dto.response.ReservationResponseDTO;
@@ -27,26 +28,29 @@ public class MakeReservationUseCase {
         this.concertService = concertService;
         this.checkTokenUseCase = checkTokenUseCase;
     }
-
+    @Transactional
     public ReservationResponseDTO execute(ReservationRequestDTO reservationRequestDTO) {
         try {
             TokenResponseDTO tokenResponseDTO = checkTokenUseCase.execute(new UserRequestDTO(reservationRequestDTO.userId()));
             String isValidToken = tokenResponseDTO.token();
 
-            Reservation checkReservation = reservationService.findNonAvailableByConcertIdAndSeatId(reservationRequestDTO.concertId(), reservationRequestDTO.seatId());
-            if(checkReservation != null){
-                return new ReservationResponseDTO("해당 좌석은 예약할 수 없습니다.", null);
-            }
             Concert concert = concertService.findByConcertId(reservationRequestDTO.concertId());
             if(concert == null) {
                 return new ReservationResponseDTO("해당 콘서트는 존재하지 않습니다.", null);
             }
             if (isValidToken != null) {
                 // 토큰이 유효한 경우, 좌석 예약 진행
+                Reservation checkReservation = reservationService.findNonAvailableByConcertIdAndSeatId(reservationRequestDTO.concertId(), reservationRequestDTO.seatId());
+                if(checkReservation != null){
+                    return new ReservationResponseDTO("해당 좌석은 예약할 수 없습니다.", null);
+                }
                 Reservation reservation = reservationService.save(new Reservation(reservationRequestDTO.userId(), reservationRequestDTO.concertId(), reservationRequestDTO.seatId(), "temporary", reservationRequestDTO.cost(), LocalDateTime.now(), LocalDateTime.now().plusMinutes(5)));
-                // 좌석 예약 후 토큰 만료처리
+                if(reservation == null) {
+                    return new ReservationResponseDTO("해당 좌석은 예약할 수 없습니다.", null);
+                }
                 tokenService.deleteByUserIdAndUseTrue(isValidToken, false);
                 return new ReservationResponseDTO("좌석 예약 성공", reservation);
+                // 좌석 예약 후 토큰 만료처리
             } else {
                 // 토큰이 유효하지 않은 경우
                 return new ReservationResponseDTO("토큰이 유효하지 않습니다.", null);
